move content user info to demo shell
