+++ demo-shell/src/app/components/app-layout/user-info/content-user-info/content-user-info.module.ts
