+++ lib/content-services/src/public-api.ts
@@ -18,7 +18,6 @@
 export * from './lib/directives/index';
 export * from './lib/tag/index';
 export * from './lib/document-list/index';
-export * from './lib/content-user-info/index';
 export * from './lib/upload/index';
 export * from './lib/search/index';
 export * from './lib/breadcrumb/index';
